ending point
I have tried for  the last 5 to 6 hours to fix this on click event. I have tried using document on ready. I have tried targeting a unique id that i put on during the render of the page. I have tried it with forms. I have tried just sending the id as a  query parameter. I cannot for the life of my understand why this click event isnt firing. I know that jquery is there because I use it to post new burgers which works fine. I cannot fathom why the click event for making a new burger works and the one for updating a burger does not. If it had something to do with being dynamically  generated then why would a document on ready not solve it? nor a click event on the entire document which then checks if its a certain class. I dont think I can focus hard enough to keep banging my head against this wall. just gonna ask for help in class tomorrow